indent by spaces

+++ OSWifiHeatMap/ViewController.swift
@@ -12,42 +12,42 @@ import DTMHeatmap
 
 class ViewController: UIViewController, MKMapViewDelegate {
 
-	@IBOutlet weak var mapView: MKMapView!
-	
-	var locations: [NSValue: Float] = [:]
-	
-	lazy var heatMap: DTMHeatmap = {
-		return DTMHeatmap()
-	}()
-	
-	override func viewDidLoad() {
-		super.viewDidLoad()
-		
-		mapView.showsUserLocation = true
-		mapView.addOverlay(heatMap)
-	}
-	
-	// MARK: - Map View Delegate
-
-	func mapView(mapView: MKMapView, didUpdateUserLocation userLocation: MKUserLocation) {
-		if let location = userLocation.location {
-			
-			if locations.count == 0 {
-				mapView.region = MKCoordinateRegionMakeWithDistance(location.coordinate, 100, 100)
-			}
-			
-			let mapPoint = MKMapPointForCoordinate(location.coordinate)
-			let value = NSValue(MKMapPoint: mapPoint)
-			locations[value] = 1
-			
-			// FIXME: force rerendering instead of readding whole overlay
-			mapView.removeOverlay(heatMap)
-			heatMap.setData(locations)
-			mapView.addOverlay(heatMap)
-		}
-	}
-	
-	func mapView(mapView: MKMapView, rendererForOverlay overlay: MKOverlay) -> MKOverlayRenderer {
-		return DTMHeatmapRenderer(overlay: overlay)
-	}
+    @IBOutlet weak var mapView: MKMapView!
+    
+    var locations: [NSValue: Double] = [:]
+    
+    lazy var heatMap: DTMHeatmap = {
+        return DTMHeatmap()
+    }()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        mapView.showsUserLocation = true
+        mapView.addOverlay(heatMap)
+    }
+    
+    // MARK: - Map View Delegate
+    
+    func mapView(mapView: MKMapView, didUpdateUserLocation userLocation: MKUserLocation) {
+        if let location = userLocation.location {
+            
+            if locations.count == 0 {
+                mapView.region = MKCoordinateRegionMakeWithDistance(location.coordinate, 100, 100)
+            }
+            
+            let mapPoint = MKMapPointForCoordinate(location.coordinate)
+            let value = NSValue(MKMapPoint: mapPoint)
+            locations[value] = 1
+            
+            // FIXME: force rerendering instead of readding whole overlay
+            mapView.removeOverlay(heatMap)
+            heatMap.setData(locations)
+            mapView.addOverlay(heatMap)
+        }
+    }
+    
+    func mapView(mapView: MKMapView, rendererForOverlay overlay: MKOverlay) -> MKOverlayRenderer {
+        return DTMHeatmapRenderer(overlay: overlay)
+    }
 }
